Use a parcel_filesystem crate
* Move the `parcel_resolver::fs` module into its own crate
* Move our in-memory and JSDelegate implementations into that crate
* Move the napi helper functions from node-bindings into another crate
* Change all usages to use these functions
