+++ crates/macros/Cargo.toml
@@ -19,5 +19,5 @@ swc_core = { version = "0.89.6", features = [
 ] }
 serde = "1.0.123"
 napi-derive = { version = "2.12.5", optional = true }
-napi = { version =  "2.12.6", features = ["serde-json", "napi4", "napi5"], optional = true }
+napi = { version =  "2.16.4", features = ["serde-json", "napi4", "napi5"], optional = true }
 crossbeam-channel = { version = "0.5.6", optional = true }
