+++ crates/node-bindings/src/core/mod.rs
@@ -1,5 +1,4 @@
 //! Core re-implementation in Rust
 
-mod filesystem;
 /// Request types and run functions
 mod requests;
