+++ crates/node-bindings/src/core/requests/config_request/mod.rs
@@ -155,10 +155,10 @@ struct RequestOptions {}
 
 #[cfg(test)]
 mod test {
-  use parcel_resolver::OsFileSystem;
+  use parcel_filesystem::in_memory_file_system::InMemoryFileSystem;
+  use parcel_filesystem::os_file_system::OsFileSystem;
 
   use super::*;
-  use crate::core::filesystem::in_memory_file_system::InMemoryFileSystem;
   use crate::core::requests::config_request::run_config_request;
   use crate::core::requests::request_api::MockRequestApi;
