+++ crates/parcel_filesystem/src/in_memory_file_system/mod.rs
@@ -3,7 +3,7 @@ use std::path::Component;
 use std::path::Path;
 use std::path::PathBuf;
 
-use parcel_resolver::FileSystem;
+use crate::FileSystem;
 
 /// In memory implementation of a file-system entry
 enum InMemoryFileSystemEntry {
@@ -49,11 +49,11 @@ impl Default for InMemoryFileSystem {
 }
 
 impl FileSystem for InMemoryFileSystem {
-  fn canonicalize<P: AsRef<Path>>(
-    &self,
-    path: P,
-    _cache: &dashmap::DashMap<PathBuf, Option<PathBuf>>,
-  ) -> std::io::Result<PathBuf> {
+  fn cwd(&self) -> std::io::Result<PathBuf> {
+    Ok(self.current_working_directory.clone())
+  }
+
+  fn canonicalize_base<P: AsRef<Path>>(&self, path: P) -> std::io::Result<PathBuf> {
     let path = path.as_ref();
 
     let mut result = if path.is_absolute() {
