+++ crates/parcel_filesystem/src/js_delegate_file_system/mod.rs
@@ -2,13 +2,13 @@ use std::path::Path;
 use std::path::PathBuf;
 use std::rc::Rc;
 
-use dashmap::DashMap;
 use napi::bindgen_prelude::FromNapiValue;
 use napi::Env;
 use napi::JsObject;
-use parcel_resolver::FileSystem;
+use napi::JsUnknown;
+use parcel_napi_helpers::call_method;
 
-use crate::core::requests::call_method;
+use crate::FileSystem;
 
 /// An implementation of `FileSystem` that delegates calls to a `JsObject`.
 ///
@@ -25,6 +25,12 @@ impl JSDelegateFileSystem {
   }
 }
 
+fn path_from_js(result: JsUnknown) -> napi::Result<PathBuf> {
+  let result_string = result.coerce_to_string()?;
+  let result_string = result_string.into_utf8()?.as_str()?.to_string();
+  Ok(PathBuf::from(result_string))
+}
+
 // Convert arbitrary errors to io errors. This is wrong; the `FileSystem` trait should use
 // `anyhow::Result`
 fn run_with_errors<T>(block: impl FnOnce() -> anyhow::Result<T>) -> Result<T, std::io::Error> {
@@ -33,11 +39,14 @@ fn run_with_errors<T>(block: impl FnOnce() -> anyhow::Result<T>) -> Result<T, st
 }
 
 impl FileSystem for JSDelegateFileSystem {
-  fn canonicalize<P: AsRef<Path>>(
-    &self,
-    path: P,
-    _cache: &DashMap<PathBuf, Option<PathBuf>>,
-  ) -> std::io::Result<PathBuf> {
+  fn cwd(&self) -> std::io::Result<PathBuf> {
+    run_with_errors(|| {
+      let result = call_method(&self.env, &self.js_delegate, "cwd", &[])?;
+      Ok(path_from_js(result)?)
+    })
+  }
+
+  fn canonicalize_base<P: AsRef<Path>>(&self, path: P) -> std::io::Result<PathBuf> {
     run_with_errors(|| {
       let path = path.as_ref().to_str().unwrap();
       let js_path = self.env.create_string(path)?;
@@ -47,9 +56,7 @@ impl FileSystem for JSDelegateFileSystem {
         "canonicalize",
         &[&js_path.into_unknown()],
       )?;
-      let result_string = result.coerce_to_string()?;
-      let result_string = result_string.into_utf8()?.as_str()?.to_string();
-      Ok(PathBuf::from(result_string))
+      Ok(path_from_js(result)?)
     })
   }
