+++ crates/parcel_napi_helpers/src/lib.rs
@@ -1,20 +1,14 @@
-use std::rc::Rc;
-
 use napi::bindgen_prelude::FromNapiValue;
 use napi::Env;
 use napi::JsFunction;
 use napi::JsObject;
-use napi::JsString;
 use napi::JsUnknown;
 use napi::NapiRaw;
-use napi_derive::napi;
 
-use crate::core::requests::config_request::ConfigRequest;
-use crate::core::requests::request_api::js_request_api::JSRequestApi;
-use parcel_filesystem::js_delegate_file_system::JSDelegateFileSystem;
-
-mod config_request;
-mod request_api;
+/// Convert anyhow error to napi error
+pub fn anyhow_napi(value: anyhow::Error) -> napi::Error {
+  napi::Error::from_reason(format!("[napi] {}", value.to_string()))
+}
 
 /// Get an object field as a JSFunction. Will error out if the field is not present or isn't an
 /// instance of the global `"Function"`.
@@ -71,41 +65,4 @@ pub fn call_method(
   let method_fn = get_function(env, js_object, field_name)?;
   let result = method_fn.call(Some(&js_object), &args)?;
   Ok(result)
-}
-
-/// JavaScript API for running a config request.
-/// At the moment the request fields themselves will be copied on call.
-///
-/// This is not efficient but can be worked around when it becomes an issue.
-///
-/// This should have exhaustive unit-tests on `packages/core/core/test/requests/ConfigRequest.test.js`.
-#[napi]
-fn napi_run_config_request(
-  env: Env,
-  config_request: ConfigRequest,
-  api: JsObject,
-  options: JsObject,
-) -> napi::Result<()> {
-  // Technically we could move `env` to JSRequestAPI but in order to
-  // be able to use env on more places we rc it.
-  let env = Rc::new(env);
-  let api = JSRequestApi::new(env.clone(), api);
-  let input_fs = options.get("inputFS")?;
-  let Some(input_fs) = input_fs.map(|input_fs| JSDelegateFileSystem::new(env, input_fs)) else {
-    // We need to make the `FileSystem` trait object-safe so we can use dynamic
-    // dispatch.
-    return Err(napi::Error::from_reason(
-      "[napi] Missing required inputFS options field",
-    ));
-  };
-  let Some(project_root): Option<JsString> = options.get("projectRoot")? else {
-    return Err(napi::Error::from_reason(
-      "[napi] Missing required projectRoot options field",
-    ));
-  };
-  // TODO: what if the string is UTF16 or latin?
-  let project_root = project_root.into_utf8()?;
-  let project_root = project_root.as_str()?;
-
-  config_request::run_config_request(&config_request, &api, &input_fs, project_root)
 }
