+++ packages/utils/node-resolver-rs/Cargo.toml
@@ -14,6 +14,7 @@ bitflags = "1.3.2"
 indexmap = { version = "1.9.2", features = ["serde"] }
 itertools = "0.10.5"
 json_comments = { path = "../../../crates/json-comments-rs" }
+parcel_filesystem = { path = "../../../crates/parcel_filesystem" }
 typed-arena = "2.0.2"
 elsa = "1.7.0"
 once_cell = "1.17.0"
